+++ packages/apps/shopify-app-express/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 6.0.4
+
+### Patch Changes
+
+- Updated dependencies [a6a13bf]
+- Updated dependencies [f1af47e]
+  - @shopify/shopify-api@12.2.0
+  - @shopify/shopify-app-session-storage@4.0.4
+  - @shopify/shopify-app-session-storage-memory@5.0.4
+
 ## 6.0.3
 
 ### Patch Changes